+++ confiture-web-app/src/components/account/dashboard/AuditRow.vue
@@ -13,6 +13,7 @@ import {
   getCriteriaCount,
   slugify
 } from "../../../utils";
+import AuditProgressBar from "../../audit/AuditProgressBar.vue";
 import DeleteModal from "../../audit/DeleteModal.vue";
 import DuplicateModal from "../../audit/DuplicateModal.vue";
 import CopyIcon from "../../icons/CopyIcon.vue";
@@ -167,17 +168,6 @@ function copyStatementLink(uniqueId: string) {
       <strong>{{ audit.procedureName }}</strong>
     </RouterLink>
 
-    <!-- Status -->
-    <p
-      class="fr-badge fr-badge--sm audit-status"
-      :class="{
-        'fr-badge--purple-glycine': isInProgress || isNotStarted
-      }"
-    >
-      <span class="fr-sr-only">Statut </span
-      >{{ isInProgress || isNotStarted ? "En cours" : "Terminé" }}
-    </p>
-
     <!-- Creation date -->
     <p class="fr-mb-0 audit-date">
       <span class="fr-sr-only">Date de création </span>
@@ -192,8 +182,8 @@ function copyStatementLink(uniqueId: string) {
       {{ getCriteriaCount(audit.auditType) }} critères
     </p>
 
-    <!-- Compliance level -->
-    <div class="audit-compliance-level">
+    <!-- Compliance level / Progression level -->
+    <div v-if="audit.status === AuditStatus.COMPLETED" class="fr-mr-4w">
       <template v-if="audit.auditType === AuditType.FULL">
         <p
           class="fr-badge fr-badge--sm fr-badge--no-icon fr-mb-0"
@@ -247,6 +237,15 @@ function copyStatementLink(uniqueId: string) {
         </span>
       </p>
     </div>
+
+    <AuditProgressBar
+      v-else
+      class="fr-mr-4w"
+      label="Progression de l’audit"
+      :value="audit.progress"
+      :size="8"
+      inline
+    />
 
     <!-- Main action -->
     <RouterLink
@@ -447,7 +446,7 @@ function copyStatementLink(uniqueId: string) {
 <style scoped>
 .grid {
   display: grid;
-  grid-template-columns: 2fr 0.75fr 0.75fr 0.75fr 1.25fr 1.5fr 1fr;
+  grid-template-columns: 2fr 0.75fr 0.75fr 1.25fr 1.5fr 1fr;
   grid-gap: 1rem;
   align-items: center;
   border: 1px solid var(--border-default-grey);