+++ confiture-web-app/src/components/audit/AuditGenerationHeader.vue
@@ -246,7 +246,9 @@ onMounted(() => {
     >
       <AuditProgressBar
         v-if="showAuditProgressBar"
+        :value="resultStore.auditProgress"
         label="Progression de l’audit"
+        :size="4"
         class="progress-bar"
       />